+++ CHANGELOG.md
@@ -27,6 +27,9 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 - A new artifact creation method for charts that accepts plotly figures as input and stores them into a json file that
   can be read by the front-end to be
   rendered ([#124](https://gitlab.heigit.org/climate-action/climatoology/-/issues/124))
+- A new 'safe' artifact creation context manager which can be used to catch errors safely while generating artifacts,
+  capturing the error messages in the computation info and as
+  warnings ([#150](https://gitlab.heigit.org/climate-action/climatoology/-/issues/150))
 - Added `expires` as an input to `get_icon_url()` to make this configurable by the API Gateway
 
 ## [6.1.2](https://gitlab.heigit.org/climate-action/climatoology/-/releases/6.1.2) - 2025-02-06